+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bumps `eslint-config-prettier` from 8.8.0 to 8.9.0
 - Bumps `sass` from 1.64.1 to 1.64.2
 - Bumps `prosemirror-view` from 1.31.6 to 1.31.7
+- Bumps `@appuniversum/ember-appuniversum` from 2.7.0 to 2.10.0
 
 ## [4.2.0] - 2023-07-29
 ### Dependencies