+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Fixed
+- All Paragraphs are now part of the group `paragraphGroup`
+  - A list will accept any `paragraphGroup`
+- ParagraphWithConfig will get parsed as itself by using the `name` attribute in the config, which gets added as an attribute. Before all paragraphs would get parsed as a basic paragraph.
 ### Dependencies
 - Bumps `eslint-config-prettier` from 8.8.0 to 8.9.0
 - Bumps `sass` from 1.64.1 to 1.64.2