+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bumps `sass` from 1.64.1 to 1.64.2
 - Bumps `prosemirror-view` from 1.31.6 to 1.31.7
 - Bumps `@types/sinon` from 10.0.15 to 10.0.16
+- Bumps `eslint` from 8.45.0 to 8.46.0
 
 ## [4.2.0] - 2023-07-29
 ### Dependencies