+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bumps `prosemirror-view` from 1.31.6 to 1.31.7
 - Bumps `@types/sinon` from 10.0.15 to 10.0.16
 - Bumps `eslint` from 8.45.0 to 8.46.0
+- Bumps `handlebars` from 4.7.7 to 4.7.8
 
 ## [4.2.0] - 2023-07-29
 ### Dependencies