+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bumps `handlebars` from 4.7.7 to 4.7.8
 - Bumps `prettier` from 3.0.0 to 3.0.1
 - Bumps `prosemirror-transform` from 1.7.3 to 1.7.4
+- Bumps `@codemirror/view` from 6.15.3 to 6.16.0
 
 ## [4.2.0] - 2023-07-29
 ### Dependencies