+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Dependencies
 - Bumps `@typescript-eslint/eslint-plugin` from 6.2.0 to 6.3.0
 - Bumps `xml-formatter` from 3.4.1 to 3.5.0
+- Bumps `eslint-config-prettier` from 8.9.0 to 9.0.0
   - A list will accept any `paragraphGroup`
 - ParagraphWithConfig accepts a config option `subType` which is required.
   - For a normal paragraph this can be the empty string